data/requirements.txt
addict==2.4.0
ansi2html==1.8.0
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
ConfigArgParse==1.5.5
contourpy==1.1.0
cycler==0.11.0
dash==2.11.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.5
fonttools==4.40.0
idna==3.4
ipykernel==6.23.3
ipython==8.14.0
ipywidgets==8.0.6
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.0
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
nbformat==5.7.0
nest-asyncio==1.5.6
numpy==1.25.0
open3d==0.17.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
plotly==5.15.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.1.0
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
retrying==1.3.4
scikit-learn==1.2.2
scipy==1.11.0
six==1.16.0
stack-data==0.6.2
tenacity==8.2.2
threadpoolctl==3.1.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.7
